requirements.txt
bottle==0.12.9
lxml==3.6.0
python-dateutil==2.5.1
requests==2.9.1
six==1.10.0
